Change multi-context support to use new login processor add_contexts parameter
The login processor in Revolution v2.0.4 and later supports a new add_contexts parameter, which specify additional contexts to log the user into. These can be supplied as contexts. Note however that a primary loginContext is still required, but can be specified either as a snippet property or a login_context $_REQUEST parameter. The contexts property of the Login snippet should not include the primary loginContext. The property can be overridden via an add_contexts $_REQUEST parameter.

To use via the add_contexts $_REQUEST parameter, add a hidden input field with the add_contexts to your login form, and an additional GET parameter on the logout form, specifying the additional contexts.

Use via the context snippet property does not require any special $_REQUEST parameters. This usage is recommended.